+++ src/pyworldatlas/__init__.py
@@ -6,13 +6,14 @@ from .exceptions import (AmbiguousCountryError, AmbiguousPlaceError, AtlasClosed
                          AtlasError, CapitalNotFoundError, CountryNotFoundError,
                          DatasetError, DatasetIntegrityError, DatasetNotFoundError,
                          DatasetVersionError, PlaceNotFoundError)
-from .models import (Area, Capital, City, Coordinate, Country, CountryCodes,
-                     CountryDiscoveryCard, CountryMatch, CountryReference,
+from .models import (Area, BorderPathResult, Capital, City, Coordinate, Country,
+                     CountryCodes, CountryDiscoveryCard, CountryMatch, CountryReference,
                      CountryStatus, Currency, DatasetInfo, Flashcard, Geography,
                      Language, LocalizedName, SourceReference)
 
 __all__ = [
     "Atlas", "Country", "CountryCodes", "CountryReference", "CountryDiscoveryCard",
+    "BorderPathResult",
     "CountryStatus", "LocalizedName", "Flashcard",
     "Coordinate", "Area", "Geography", "Capital", "City", "Currency", "Language", "CountryMatch",
     "DatasetInfo", "SourceReference", "AtlasError", "AtlasClosedError",